adding utils service for GUID
